refactor: update uriMatchType filename
