+++ libs/common/src/vault/models/data/login-uri.data.ts
@@ -1,4 +1,4 @@
-import { UriMatchType } from "../../../enums/uriMatchType";
+import { UriMatchType } from "../../../enums/uri-match-type";
 import { LoginUriApi } from "../../../models/api/login-uri.api";
 
 export class LoginUriData {
