+++ libs/common/src/vault/models/domain/login-uri.spec.ts
@@ -1,7 +1,7 @@
 import { Jsonify } from "type-fest";
 
 import { mockEnc, mockFromJson } from "../../../../spec/utils";
-import { UriMatchType } from "../../../enums/uriMatchType";
+import { UriMatchType } from "../../../enums/uri-match-type";
 import { EncString } from "../../../models/domain/enc-string";
 import { LoginUriData } from "../data/login-uri.data";
