+++ libs/common/src/vault/models/domain/login-uri.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { UriMatchType } from "../../../enums/uriMatchType";
+import { UriMatchType } from "../../../enums/uri-match-type";
 import Domain from "../../../models/domain/domain-base";
 import { EncString } from "../../../models/domain/enc-string";
 import { SymmetricCryptoKey } from "../../../models/domain/symmetric-crypto-key";
