+++ libs/common/src/vault/models/view/login-uri.view.ts
@@ -1,6 +1,6 @@
 import { Jsonify } from "type-fest";
 
-import { UriMatchType } from "../../../enums/uriMatchType";
+import { UriMatchType } from "../../../enums/uri-match-type";
 import { Utils } from "../../../misc/utils";
 import { View } from "../../../models/view/view";
 import { LoginUri } from "../domain/login-uri";
